pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bob"
version = "0.1.0"
description = "B.O.B – Boring Organizer Bot: A local-first personal knowledge assistant"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "B.O.B Contributors" }
]
keywords = ["knowledge-management", "local-first", "embeddings", "search", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Indexing",
]

dependencies = [
    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",
    
    # Configuration
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Database
    "sqlite-vec>=0.1.0",
    
    # Embeddings
    "sentence-transformers>=2.2.0",
    
    # Document parsing
    "markdown-it-py>=3.0.0",
    "pypdf>=3.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    
    # Git operations
    "gitpython>=3.1.0",
    
    # Utilities
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "types-PyYAML>=6.0.0",
    "types-python-dateutil>=2.8.0",
]
llm = [
    # Optional local LLM support
    "llama-cpp-python>=0.2.0",
]

[project.scripts]
bob = "bob.cli.main:cli"

[project.urls]
Homepage = "https://github.com/sandgraal/Boring-Organizer-Bot"
Documentation = "https://github.com/sandgraal/Boring-Organizer-Bot/tree/main/docs"
Repository = "https://github.com/sandgraal/Boring-Organizer-Bot"
Issues = "https://github.com/sandgraal/Boring-Organizer-Bot/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["bob*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["bob"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "sqlite_vec",
    "sentence_transformers",
    "pypdf",
    "docx",
    "git",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["bob"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
